Enhance FilesLister
This new version of FilesLister relies in the path parsing mechanism
that is used by Spark. Globbing patterns are now supported, with the
only condition that they must directly reference the files that need to
be used as input files, and not the directories they are in. For
example:

```
animals/
cats/
    cat1.txt
    cat2.txt
dogs/
    dog1.txt
```

The expression:

```
   animals/*
```

would end up being invalid, as no `.txt` files are referenced, but only
directories (`cats` and `dogs`). This can be done doing:

```
   animals/*/*
```

Future versions of this library might change this behaviour.